fix error handling

+++ Set.py
@@ -9,6 +9,7 @@ def add_set(set_id: int):
     except sqlite3.Error as e:
         print('Error connecting to database')
         print(e)
+        return
 
     c = conn.cursor()
     try:
@@ -16,10 +17,10 @@ def add_set(set_id: int):
             INSERT INTO sets (set_id)
             VALUES ({set_id});
         """)
+
+        conn.commit()
     except sqlite3.IntegrityError as e:
         print(f'Set {set_id} is already in database')
-        print(e)
-        conn.commit()
     finally:
         conn.close()
 
@@ -31,20 +32,51 @@ def delete_set(set_id: int):
     except sqlite3.Error as e:
         print('Error connecting to database')
         print(e)
+        return
 
     c = conn.cursor()
     
     try:
         c.execute(f"""
+            SELECT * FROM sets
+            WHERE set_id = {set_id};
+        """)
+
+        if not c.fetchone():
+            raise Exception(f'Set {set_id} is not in database')
+
+        c.execute(f"""
             DELETE FROM sets
             WHERE set_id = {set_id};
         """)
-    except sqlite3.IntegrityError as e:
-        print(f'Set {set_id} is not in database')
-        print(e)
+
         conn.commit()
+    except Exception as e:
+        print(e)
     finally:
         conn.close()
+
+def get_all_sets():
+    """Returns a list of all sets in sets table"""
+
+    try:
+        conn = sqlite3.connect('./database/database.db')
+    except sqlite3.Error as e:
+        print('Error connecting to database')
+        print(e)
+        return
+
+    c = conn.cursor()
+    
+    try:
+        c.execute(f"""
+            SELECT set_id
+            FROM sets;
+        """)
+    finally:
+        sets = [set_id[0] for set_id in c.fetchall()]
+        conn.close()
+        return sets
 
 
 class Set:
